Fix two issues importing old csv format
When finding the common part of all the paths in the csv file, we
could land in the middle of a directory name, so in that case we
need to look backwards to the nearest path separator.

When the csv does not include a 'site' column, we need to use the
previous logic for extracting the site name from the nifti path.