Update for more discoveries running mriqc
Add a note indicating that the slash is needed at the end of the mriqc
path (we should fix this, but it works this way for now).

Also, I added the --no-sub argument when running mriqc so that we avoid
submission of the computed metrics back to mriqc.nimh.nih.gov, at least
while we are doing a lot of testing of our workflow.