Add a placeholder for neural network based learning
This adds a radio group to the retrain dialog, allowing the user to
select neural network or random forest classifaction.  Also adds a
placeholder for neural network learning in the tasks.py module.

Also fixes the bug where the retrain dialog "Cancel" button does
not work, and adds a note to dev doc on running "data2mriqc.py".