Fix load state indicators
Previously scans were marked as loaded or not loaded using filled
vs. empty disc next to each scan.  The computation for this was done
efficiently, but was ultimately incorrect as the scan was marked as
loaded as soon as a worker had picked up the last image to load.  In
cases where the only image in a scan is really large, the scan could
be marked as loaded before it was actually ready.

This change removes the filled circles indicating individual scans
are loaded or not in favor of a single icon located next to the
experiment name.