Fix email from address and reload logout issue
When sending email, set the from address to the email of the active
user, so that recipients can respond directly.  Also add active user
to bcc list as a chip that can be deleted.  Make sure that even if the
user reloads the browser without having to relogin, the periodic user
check is still configured to make sure user is logged out when session
cookie expires.